requirements.txt
google-cloud-storage
google-cloud-bigquery
pydicom
Pillow
numpy